refactor: change location of templates
