+++ src/db-handlers/checkKey.ts
@@ -1,5 +1,5 @@
 import * as bcrypt from 'bcrypt';
-import UserReturnData from '../routes/UserReturnData';
+import UserReturnData from '../UserReturnData';
 
 export default async (db, clientID: string, clientKey: string): Promise<boolean> => {
   const data: UserReturnData = { err_occur: false, err_reasons: [], data: {} };
