+++ src/db-handlers/fetchMatchesForCompetition.ts
@@ -1,4 +1,4 @@
-import UserReturnData from '../routes/UserReturnData';
+import UserReturnData from '../UserReturnData';
 
 export default async (db: any, competition: string): Promise<UserReturnData> => {
   const data: UserReturnData = { err_occur: false, err_reasons: [], data: { competition: undefined, data: undefined } };
