+++ src/db-handlers/fetchScouterUIDs.ts
@@ -1,4 +1,4 @@
-import UserReturnData from '../routes/UserReturnData';
+import UserReturnData from '../UserReturnData';
 
 export default async (db: any, competition: string, match: number): Promise<UserReturnData> => {
   const data: UserReturnData = {
