+++ src/db-handlers/submitPitData.ts
@@ -1,5 +1,5 @@
-import UserReturnData from '../routes/UserReturnData';
-import Scouter from '../../src/routes/Scouter';
+import UserReturnData from '../UserReturnData';
+import Scouter from '../Scouter';
 
 export default async (db: any, scouter: Scouter, competition: string, match: number, team_scouted: number, datain: Record<string, any>): Promise<UserReturnData> => {
   const data: UserReturnData = { err_occur: false, err_reasons: [], data: {} };
