+++ src/index.ts
@@ -1,7 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import expressMongoDb from 'mongo-express-req';
-
 import dbHandler = require('./dbHandler');
 import auth = require('./authHandler');
 
@@ -32,7 +31,6 @@ try {
 
 // TODO: find a way to loop through this without a bunch of require(). A simple for loop results in `require() not found`.
 // TODO: use the UserReturnData class when returning data in all these apis
-
 require('./routes/base')(app);
 require('./routes/fetchMatches')(app, dbHandler);
 require('./routes/submitMatchData')(app, dbHandler, auth);
@@ -56,7 +54,6 @@ require('./routes/fetchPitData')(app, dbHandler);
 require('./routes/submitPitData')(app, dbHandler, auth);
 require('./routes/addUserToTeam')(app, dbHandler, auth);
 require('./routes/fetchMatchConfig')(app, dbHandler);
-
 
 app.listen(port, () => console.log(`Listening on port ${port}`));
