+++ src/routes/addAPIKey.ts
@@ -1,6 +1,6 @@
 import uuidAPIKey from 'uuid-apikey';
-import UserReturnData from './UserReturnData';
-import StatusCodes from './StatusCodes';
+import UserReturnData from '../UserReturnData';
+import StatusCodes from '../StatusCodes';
 
 /**
  * POST route '/api/addAPIKey'
