+++ src/routes/fetchAllTeamNicknamesAtCompetition.ts
@@ -1,5 +1,5 @@
-import UserReturnData from './UserReturnData';
-import StatusCodes from './StatusCodes';
+import UserReturnData from '../UserReturnData';
+import StatusCodes from '../StatusCodes';
 
 /**
  * GET route '/api/fetchAllTeamNicknamesAtCompetition'
