+++ src/routes/fetchUserStrategy.ts
@@ -1,5 +1,5 @@
-import UserReturnData from './UserReturnData';
-import StatusCodes from './StatusCodes';
+import UserReturnData from '../UserReturnData';
+import StatusCodes from '../StatusCodes';
 
 module.exports = (app:any, dbHandler:any, auth: any) => {
   app.get('/api/fetchUserStrategy', auth.checkAuth, async (req: any, res:any) => {
